karrio/requirements.txt
--extra-index-url https://karrio.gateway.scarf.sh/simple/
asgiref==3.9.2
attrs==25.3.0
backoff==2.2.1
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
cssselect2==0.8.0
deprecation==2.1.0
diff-match-patch==20241021
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.6
django-constance==4.3.2
django-cors-headers==4.9.0
django-downloadview==2.4.0
django-email-verification==0.3.3
django-filter==25.1
django-formtools==2.5.1
django-health-check==3.20.0
django-import-export==4.3.10
django-oauth-toolkit==3.0.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-picklefield==3.3
django-redis==6.0.0
django-two-factor-auth==1.18.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-api-tracking==1.8.4
drf-spectacular==0.28.0
fonttools==4.60.1
googleapis-common-protos==1.70.0
graphql-core==3.2.6
grpcio==1.75.1
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
huey==2.5.3
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jsonfield==3.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jstruct==2021.11
jwcrypto==1.5.6
karrio==2025.5rc23
karrio-australiapost==2025.5rc23
karrio-canadapost==2025.5rc23
karrio-chronopost==2025.5rc23
karrio-dhl-express==2025.5rc23
karrio-dhl-parcel-de==2025.5rc23
karrio-dhl-poland==2025.5rc23
karrio-dhl-universal==2025.5rc23
karrio-dpd==2025.5rc23
karrio-easypost==2025.5rc23
karrio-easyship==2025.5rc23
karrio-eshipper==2025.5rc23
karrio-fedex==2025.5rc23
karrio-freightcom==2025.5rc23
karrio-generic==2025.5rc23
karrio-laposte==2025.5rc23
karrio-purolator==2025.5rc23
karrio-sapient==2025.5rc23
karrio-seko==2025.5rc23
karrio-sendle==2025.5rc23
karrio-server==2025.5rc23
karrio-server-core==2025.5rc23
karrio-server-data==2025.5rc23
karrio-server-documents==2025.5rc23
karrio-server-events==2025.5rc23
karrio-server-graph==2025.5rc23
karrio-server-manager==2025.5rc23
karrio-server-orders==2025.5rc23
karrio-server-pricing==2025.5rc23
karrio-server-proxy==2025.5rc23
karrio-ups==2025.5rc23
karrio-usps==2025.5rc23
karrio-usps-international==2025.5rc23
lia-web==0.2.3
lxml==4.9.4
lxml-stubs==0.5.1
MarkupSafe==3.0.3
more-itertools==10.8.0
oauthlib==3.3.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
posthog==6.7.6
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
py-soap==2020.3.2
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
PyPDF2==3.0.1
pyphen==0.17.2
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
pyzint==0.1.10
qrcode==8.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sentry-sdk==2.39.0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.282.0
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
validators==0.35.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
wrapt==1.17.3
xmltodict==1.0.2
zipp==3.23.0
zopfli==0.2.3.post1

